Create Bank Encapsulation class
This class is a Encapsulated class .Due to Encapsulation Direct accessing is not possible so we are trying to make Indirect Accessing by using getters and Setters .We  actually need unique account for all new customers and It cannot be taken from user .It should be generated by our side. For that we are initializing  a random number in a static variable called 'random'. We are actually assigning the variable to a another variable called 'account No' so whenever the the user try to create account we will increment the account no by 1.so finally we will get unique account no for all of them.